Event Timing flush pointerdown & keydown on contextmenu
* Previous undesired outcome: When right click brings up contextmenu,
the pointerdown or keydown entry is missing the first time you do so.
* Root cause: Pointerdown or keydown entries are deferred notifying PO
until pointercancel/pointerup or keyup. When pointerdown is only
followed by a contextmenu, the pointerdown/keydown entry would get stuck
in the map. The stuck entry can only be rescued when a new pointerdown
comes in to substitute it out. This explains why only the first
pointerdown was missing. It also explains why all subsequent right
clicks were printing pointerdowns, but are actually the incorrect
previous ones stuck in the map.
* This CL address it by: flushing all pointerdown, pointerup, and
keydown entries upon contextmenu.

Bug: 1413096, 1428603
Change-Id: I03e9cdbe57c146fe61e6758b39249048aa982897