FSA: Relax file name checks in the BucketFS
Unlike when saving files to the local file system, the names of files in
the the Bucket File System (a.k.a. OPFS) are obfuscated before they end
up on disk - if they even end up on disk at all.

As such, we don't need to perform the same name sanitization and
obfuscation for these files as we do for files that end up on the
user-visible file system.

See https://github.com/whatwg/fs/issues/93 for context

Bug: 1399536
Change-Id: I019b393b731cd20aa4c45eade4eca19b6633e9cd