[scroll-start] Use scroll-start after layout
scroll-start[1] should set the initial scroll position of a scroller
This is an alternative approach to crrev.com/c/4545951.

This approach does not result in scroll events being fired due to scroll
start and will also not changes the scroll offset due to changes in
overflow.
However, if a scroller is toggled from display:block to display:none and
then back to display:block, scroll-start has no effect when it becomes
display:block again. This is because the scroll offset is saved at the
element level[2].

[1] https://drafts.csswg.org/css-scroll-snap-2/#scroll-start
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/paint/paint_layer_scrollable_area.cc?q=paint_layer_scrollable_area.cc%20%22element-%3ESetSavedLayerScrollOffset(scroll_offset_)%22&ss=chromium%2Fchromium%2Fsrc

Bug: 1439807
Change-Id: I5655f8e74b85fe45038f700091065fa11eeac0f8