[anchor-position] Support anchor-scroll with fragmented containers
We currently check only the first fragment of the containing block for
anchor-scroll, and therefore can fail if the containing block is
fragmented.

This patch fixes it by iterating through all the fragments and check
every NGPhysicalAnchorQuery there. It also provides two utility
functions that are better abstracted, so that clients like
anchor-scroll and possibly DevTools (in the future) don't need to
deal with fragmentation details.

This patch also reveals an issue with relpos inline containing block,
and marks it with TODOs and a currently failing test. It will be fixed
by a followup patch.

Bug: 1446442
Change-Id: Ia08fd094616c80b40d6cbb555ae3153fbfaea04c