Remove FencedFrameConfig URL attribute
We're removing what we colloquially refer to as "default mode" in fenced
frames, which is where you manually construct a `new
FencedFrameConfig()` object with a (usually) non-opaque URL, to use
fenced frames without relying on cross-site data from FLEDGE or Shared
Storage. If a use-case that requires this mode of fenced frames presents
itself in the future, we will re-add it. But for now, this entails:
  1. Removing the FencedFrameConfig#url IDL attribute
    - Handled by this CL
    - See corresponding spec change:
      https://github.com/WICG/fenced-frame/pull/88
  2. Removing the FencedFrameConfig web-exposed constructor
    - This will happen in a subsequent CL

We're removing the URL attribute of the FencedFrameConfig interface [1] because if fenced frames can only be used with FLEDGE/Shared Storage,
the URL member is always "opaque", since the true value depends on
cross-site data. An always "opaque"-returning attribute does not add any
value to web developers.

To carry out this change we must remove tests that rely on observing
the FencedFrameConfig#url attribute, as these tests will need to be
removed anyways when fully removing "default mode".

[1]: https://wicg.github.io/fenced-frame/#fenced-frame-config-interface

R=gtanzer@chromium.org

Bug: 1123606
Change-Id: Id29e3b9dac9241d84722c3695f59d9865485d4a5