WIP: do not cache CCNS page if WebRTC/Transport/Socket used
In the current implementation, if the WebRTC/WebTransport/Websocket
connections are closed, the non-sticky feature will be removed and the
page will be eligible for BFCache.
However, the page may contain some sensitive data that are retrieved
over these channels before the connection closure, so we should not
store the CCNS page into BFCache.

This CL also create `kAllowedForNonCacheControlNoStoreFeatures` for
those features that are allowed for non-CCNS page BFCache, and moved
`kAuthorizationHeader` over.

This CL is WIP since the tests are not added yet.

Bug: 1446619
Change-Id: Ic2e1e49d50fe522b1a66c6e011cb6fa83670d05d